pic could not compute a radius, and two register-class bugs did it
`circle rad 0.5' was rejected as "invalid radius 0.000000". Both causes are
floating-point calling convention, and each hid the other -- fixing either
alone changes nothing you can see.

1. picl.l declared `extern float atof()' and atof returns a DOUBLE. On the
   VAX both came back in r0/r1 so it was harmless; on ARM64 a float return
   is s0 and a double return is d0, the same register read at a different
   width, so atof("0.5") gave 0. Measured side by side: 0.000000 against
   0.500000. A sweep found FIVE sites, all upstream's -- pic/picl.l and
   troff's ta.c, hc.c, tc.c, devi10/makefonts.c. Same family as the
   yylval.p token bug: a declaration that lies about a type.

2. The math came from Apple's libm on the wrong convention. v8cc passes
   every argument positionally in x0-x7 including doubles --

	ldr d16,[x9] ; str d16,[sp,#384] ; ldr x0,[sp,#384] ; bl _sqrt

   -- and AAPCS64 puts a double in d0, so sqrt(2.0) returned 0.000000.

tests/kmemu had carried libm as its last allowed leak with a note saying it
was "non-variadic, so it works and nothing looked wrong". That argument is
about the SHAPE of the call and is only as good as the register classes
agreeing. They do not for floating point.

There is also no libm in V8's tree to port -- the math is in libc/math,
which is why "port libm" was the wrong question and reading the tree was
the only way to find that out. All eighteen files compile under v8cc
unmodified (fabs excepted: ieeefp.o already defines it). pic and grap now
import NOTHING from the host, and the allowed-leak list is empty for the
first time.

WHY THE SUITE MISSED IT, twice over. Its pic inputs used only default sizes
-- compiled-in constants that call neither atof nor the math library -- and
asserted that drawing commands come out, not that the numbers in them are
right. And the end-to-end case counted `grep -c "^D"' on troff's device
stream. troff emits a draw preceded by the motion that positions it,
`h97Dl 0 96 .', but while every coordinate was zero there was no motion, so
each D began its line. The pattern had been calibrated against output from
a program that could not do arithmetic, and fixing pic made that test fail.
That is the second time this session a test has been found asserting the
broken behaviour.

872 tests, 16 suites, all green.

+++ src/cmd/troff/devi10/makefonts.c
@@ -931,7 +931,7 @@ newwidth(ch, width)
 	float	delta;						/* width change for the character */
 
 	float	getdelta();
-	float	atof();
+	double	atof();	/* PORT: was float; atof returns double */
 
 
 	/********************************************************************
